requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
Django==4.2.20
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg==3.2.6
# psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pytest-mock==3.14.0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-mock==1.12.1
sqlparse==0.5.3
svgwrite==1.4.3
tomli==2.2.1
Tree==0.2.4
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==1.26.20
